rustiflow::flows::basic_flow

Struct BasicFlow

pub struct BasicFlow {

    pub flow_key: String,
    pub ip_destination: IpAddr,
    pub ip_source: IpAddr,
    pub port_destination: u16,
    pub port_source: u16,
    pub protocol: u8,
    pub first_timestamp_us: i64,
    pub last_timestamp_us: i64,
    pub flow_expire_cause: FlowExpireCause,
    pub(crate) state_fwd: FlowState,
    pub(crate) state_bwd: FlowState,
    expected_ack_seq_fwd: Option<u32>,
    expected_ack_seq_bwd: Option<u32>,
    saw_syn_fwd: bool,
    saw_syn_ack_bwd: bool,
    expected_handshake_ack_seq_fwd: Option<u32>,
    pub tcp_handshake_completed: bool,
    pub tcp_reset_before_handshake: bool,
    pub tcp_reset_after_handshake: bool,
    pub tcp_close_style: TcpCloseStyle,
    saw_fin_fwd: bool,
    saw_fin_bwd: bool,
    tcp_simultaneous_close: bool,

}

A basic flow that stores the basic features of a flow.

Fields
flow_key: String

The unique identifier of the flow.

ip_destination: IpAddr

The destination IP address of the flow.

ip_source: IpAddr

The source IP address of the flow.

port_destination: u16

The destination port of the flow.

port_source: u16

The source port of the flow.

protocol: u8

The protocol of the flow.

first_timestamp_us: i64

The first timestamp of the flow.

last_timestamp_us: i64

The last timestamp of the flow.

flow_expire_cause: FlowExpireCause

The reason this flow expired

state_fwd: FlowStatestate_bwd: FlowStateexpected_ack_seq_fwd: Option<u32>expected_ack_seq_bwd: Option<u32>saw_syn_fwd: boolsaw_syn_ack_bwd: boolexpected_handshake_ack_seq_fwd: Option<u32>tcp_handshake_completed: booltcp_reset_before_handshake: booltcp_reset_after_handshake: booltcp_close_style: TcpCloseStylesaw_fin_fwd: boolsaw_fin_bwd: booltcp_simultaneous_close: bool
Implementations

impl BasicFlow

fn is_tcp(&self) -> bool

fn observe_tcp_handshake(&mut self, packet: &PacketFeatures, forward: bool)

fn update_tcp_close_style(&mut self, cause: FlowExpireCause)

pub fn is_tcp_finished( &mut self, packet: &PacketFeatures, forward: bool, ) -> bool

Checks if the flow is finished.

A flow is considered finished when both FIN flags are set and the last ACK is received, and the sequence numbers have been acknowledged by both parties.

Arguments

packet - The packet to be checked.

Returns

A boolean indicating if the flow is finished.

pub fn get_flow_duration_usec(&self) -> i64

Calculates the flow duration in microseconds.

Returns the difference between the last and first packet timestamps in microseconds.

Returns

The duration of the flow in microseconds.

pub fn get_flow_duration_msec(&self) -> i64

Calculates the flow duration in milliseconds.

Returns the difference between the last and first packet timestamps in milliseconds.

Returns

The duration of the flow in milliseconds.

pub fn get_last_timestamp(&self) -> DateTime<Utc>

pub fn get_first_timestamp(&self) -> DateTime<Utc>

pub fn get_ip_version(&self) -> u8

pub fn get_source_ip_scope(&self) -> IpScope

pub fn get_destination_ip_scope(&self) -> IpScope

pub fn get_path_locality(&self) -> PathLocality