rustiflow::flows::features::util

Trait FlowFeature

pub trait FlowFeature:
    Send
    + Sync
    + Clone {
    // Required methods
    fn update(
        &mut self,
        packet: &PacketFeatures,
        is_forward: bool,
        last_timestamp: i64,
    );
    fn close(&mut self, last_timestamp: i64, cause: FlowExpireCause);
    fn dump(&self) -> String;
    fn headers() -> String
       where Self: Sized;
}

Trait for network flow features that can be updated, closed, and dumped to CSV format

Required Methods

fn update( &mut self, packet: &PacketFeatures, is_forward: bool, last_timestamp: i64, )

Updates the feature with a new packet

fn close(&mut self, last_timestamp: i64, cause: FlowExpireCause)

Finalizes any active state when the flow is terminated

fn dump(&self) -> String

Dumps the current state as a CSV string

fn headers() -> String
where Self: Sized,

Returns the CSV headers for this feature

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.