rustiflow::flows::ntl_flow
Struct NTLFlow

pub struct NTLFlow {
    pub cic_flow: CicFlow,
    fwd_header_len_min: u32,
    fwd_header_len_max: u32,
    fwd_header_len_mean: f32,
    fwd_header_len_std: f32,
    bwd_header_len_min: u32,
    bwd_header_len_max: u32,
    bwd_header_len_mean: f32,
    bwd_header_len_std: f32,
}

Represents a NTL Flow, encapsulating various metrics and states of a network flow.

This flow represents the same flow as the NTLFlowLyzer does.

Fields
cic_flow: CicFlow

The cic flow information.

fwd_header_len_min: u32

The minimum header length of the forward flow.

fwd_header_len_max: u32

The maximum header length of the forward flow.

fwd_header_len_mean: f32

The mean packet length of the forward flow.

fwd_header_len_std: f32

The std header length of the forward flow.

bwd_header_len_min: u32

The minimum header length of the backward flow.

bwd_header_len_max: u32

The maximum header length of the backward flow.

bwd_header_len_mean: f32

The mean packet length of the backward flow.

bwd_header_len_std: f32

The std header length of the backward flow.

Implementations

impl NTLFlow

fn update_fwd_header_len_stats(&mut self, len: u32)

Updates statistics for the header length of forward packets.

This method updates the maximum, minimum, total, mean, and standard deviation for the header lengths of forward packets. Itโs used for analyzing header size distributions in the forward direction of the flow.

Arguments

len - The length of the new packet header to be incorporated into the statistics.

fn update_bwd_header_len_stats(&mut self, len: u32)

Updates statistics for the length of forward packets.

This method updates the maximum, minimum, total, mean, and standard deviation for the lengths of forward packets. Itโs used for analyzing packet size distributions in the forward direction of the flow.

Arguments

len - The length of the new packet to be incorporated into the statistics.

pub fn get_fwd_header_length_min(&self) -> u32

Retrieves the minimum packet header length for forward packets.

Returns the minimum packet header length observed in the forward direction. Returns 0 if the minimum header length is not set (indicated by u32::MAX).

Returns

Minimum forward packet header length, or 0 if not set.

pub fn get_bwd_header_length_min(&self) -> u32

Retrieves the minimum packet header length for backward packets.

Returns the minimum packet header length observed in the backward direction. Returns 0 if the minimum header length is not set (indicated by u32::MAX).

Returns

Minimum backward packet header length, or 0 if not set.

pub fn get_flow_header_length_max(&self) -> u32

Retrieves the maximum packet header length in the flow, considering both forward and backward directions.

Compares the maximum packet header lengths of forward and backward flows and returns the larger value.

Returns

Maximum packet header length in the flow.

pub fn get_flow_header_length_min(&self) -> u32

Retrieves the minimum packet header length in the flow, considering both forward and backward directions.

Compares the minimum packet header lengths of forward and backward flows and returns the smaller value. Returns 0 if the minimum length is not set (indicated by u32::MAX).

Returns

Minimum packet header length in the flow, or 0 if not set.

pub fn get_flow_header_length_mean(&self) -> f32

Calculates the mean packet header length of the flow, averaging both forward and backward packet header lengths.

The mean is computed by considering the header lengths and counts of packets in both directions.

Returns

Mean packet header length of the flow.

pub fn get_flow_header_length_variance(&self) -> f64

Calculates the variance of the packet lengths in the flow.

Computes the variance by considering the standard deviations of packet lengths in both forward and backward directions.

Returns

Variance of the flowโs packet lengths, or 0 if not enough data.

pub fn get_flow_header_length_std(&self) -> f64

Retrieves the standard deviation of packet lengths in the flow.

Utilizes the calculated variance of packet lengths to compute the standard deviation.

Returns

Standard deviation of the flowโs packet lengths.