rustiflow

Module realtime

Structs

PacketGraphState 🔒

Functions

bump_memlock_rlimit 🔒

compute_realtime_offset_us 🔒

compute_shard_index 🔒

ebpf_binary_path 🔒

handle_realtime

Starts the realtime processing of packets on the given interface. The function will return the number of packets dropped by the eBPF program.

load_ebpf_ipv4 🔒

load_ebpf_ipv6 🔒

read_clock_us 🔒