rustiflow::realtime

Function handle_realtime

pub async fn handle_realtime<T>(
    interface: &str,
    output_channel: Sender<T>,
    num_threads: u8,
    active_timeout: u64,
    idle_timeout: u64,
    early_export: Option<u64>,
    expiration_check_interval: u64,
    ingress_only: bool,
    packet_graph_mode: PacketGraphMode,
) -> Result<u64, Error>
where
    T: Flow,

Starts the realtime processing of packets on the given interface. The function will return the number of packets dropped by the eBPF program.