rustiflow::flows::basic_flow
Struct BasicFlow

pub struct BasicFlow {

    pub flow_key: String,
    pub ip_destination: IpAddr,
    pub ip_source: IpAddr,
    pub port_destination: u16,
    pub port_source: u16,
    pub protocol: u8,
    pub first_timestamp: DateTime<Utc>,
    pub last_timestamp: DateTime<Utc>,
    pub flow_end_of_flow_ack: u8,
    pub fwd_fin_flag_count: u32,
    pub fwd_syn_flag_count: u32,
    pub fwd_rst_flag_count: u32,
    pub fwd_psh_flag_count: u32,
    pub fwd_ack_flag_count: u32,
    pub fwd_urg_flag_count: u32,
    pub fwd_cwe_flag_count: u32,
    pub fwd_ece_flag_count: u32,
    pub fwd_packet_count: u32,
    pub bwd_fin_flag_count: u32,
    pub bwd_syn_flag_count: u32,
    pub bwd_rst_flag_count: u32,
    pub bwd_psh_flag_count: u32,
    pub bwd_ack_flag_count: u32,
    pub bwd_urg_flag_count: u32,
    pub bwd_cwe_flag_count: u32,
    pub bwd_ece_flag_count: u32,
    pub bwd_packet_count: u32,
    state_fwd: FlowState,
    state_bwd: FlowState,
    expected_ack_seq_fwd: Option<u32>,
    expected_ack_seq_bwd: Option<u32>,

}

A basic flow that stores the basic features of a flow.

Fields
flow_key: String

The unique identifier of the flow.

ip_destination: IpAddr

The destination IP address of the flow.

ip_source: IpAddr

The source IP address of the flow.

port_destination: u16

The destination port of the flow.

port_source: u16

The source port of the flow.

protocol: u8

The protocol of the flow.

first_timestamp: DateTime<Utc>

The first timestamp of the flow.

last_timestamp: DateTime<Utc>

The last timestamp of the flow.

flow_end_of_flow_ack: u8

The last ACK of the flow.

fwd_fin_flag_count: u32

The number of FIN flags in the forward direction.

fwd_syn_flag_count: u32

The number of SYN flags in the forward direction.

fwd_rst_flag_count: u32

The number of RST flags in the forward direction.

fwd_psh_flag_count: u32

The number of PSH flags in the forward direction.

fwd_ack_flag_count: u32

The number of ACK flags in the forward direction.

fwd_urg_flag_count: u32

The number of URG flags in the forward direction.

fwd_cwe_flag_count: u32

The number of CWE flags in the forward direction.

fwd_ece_flag_count: u32

The number of ECE flags in the forward direction.

fwd_packet_count: u32

The number of packets in the forward direction.

bwd_fin_flag_count: u32

The number of FIN flags in the backward direction.

bwd_syn_flag_count: u32

The number of SYN flags in the backward direction.

bwd_rst_flag_count: u32

The number of RST flags in the backward direction.

bwd_psh_flag_count: u32

The number of PSH flags in the backward direction.

bwd_ack_flag_count: u32

The number of ACK flags in the backward direction.

bwd_urg_flag_count: u32

The number of URG flags in the backward direction.

bwd_cwe_flag_count: u32

The number of CWE flags in the backward direction.

bwd_ece_flag_count: u32

The number of ECE flags in the backward direction.

bwd_packet_count: u32

The number of packets in the backward direction.

state_fwd: FlowStatestate_bwd: FlowStateexpected_ack_seq_fwd: Option<u32>expected_ack_seq_bwd: Option<u32>
Implementations

impl BasicFlow

pub fn is_tcp_finished( &mut self, packet: &PacketFeatures, forward: bool, ) -> bool

Checks if the flow is finished.

A flow is considered finished when both FIN flags are set and the last ACK is received, and the sequence numbers have been acknowledged by both parties.

Arguments

packet - The packet to be checked.

Returns

A boolean indicating if the flow is finished.

pub fn get_flow_duration_usec(&self) -> f64

Calculates the flow duration in microseconds.

Returns the difference between the last and first packet timestamps in microseconds.

Returns

The duration of the flow in microseconds.