Struct rustiflow::flows::cic_flow::CicFlow

pub struct CicFlow {

    pub basic_flow: BasicFlow,
    pub sf_last_packet_timestamp: Option<Instant>,
    pub sf_count: u32,
    pub start_active: Instant,
    pub end_active: Instant,
    pub active_count: u32,
    pub active_mean: f64,
    pub active_std: f64,
    pub active_max: f64,
    active_min: f64,
    pub idle_count: u32,
    pub idle_mean: f64,
    pub idle_std: f64,
    pub idle_max: f64,
    idle_min: f64,
    pub fwd_init_win_bytes: u16,
    pub fwd_act_data_pkt: u32,
    fwd_header_len_min: u32,
    pub fwd_last_timestamp: Option<Instant>,
    pub fwd_pkt_len_tot: u32,
    pub fwd_seg_len_tot: u32,
    pub fwd_pkt_len_max: u32,
    fwd_pkt_len_min: u32,
    pub fwd_pkt_len_mean: f32,
    pub fwd_pkt_len_std: f32,
    pub fwd_iat_total: f64,
    pub fwd_iat_mean: f64,
    pub fwd_iat_std: f64,
    pub fwd_iat_max: f64,
    fwd_iat_min: f64,
    pub fwd_header_length: u32,
    pub fwd_bulk_duration: f64,
    pub fwd_bulk_packet_count: u64,
    pub fwd_bulk_size_total: u32,
    pub fwd_bulk_state_count: u64,
    fwd_bulk_packet_count_help: u64,
    fwd_bulk_start_help: Option<Instant>,
    fwd_bulk_size_help: u32,
    pub fwd_last_bulk_timestamp: Option<Instant>,
    pub bwd_init_win_bytes: u16,
    pub bwd_last_timestamp: Option<Instant>,
    pub bwd_pkt_len_tot: u32,
    pub bwd_seg_len_tot: u32,
    pub bwd_pkt_len_max: u32,
    bwd_pkt_len_min: u32,
    pub bwd_pkt_len_mean: f32,
    pub bwd_pkt_len_std: f32,
    pub bwd_iat_total: f64,
    pub bwd_iat_mean: f64,
    pub bwd_iat_std: f64,
    pub bwd_iat_max: f64,
    bwd_iat_min: f64,
    pub bwd_header_length: u32,
    pub bwd_bulk_duration: f64,
    pub bwd_bulk_packet_count: u64,
    pub bwd_bulk_size_total: u32,
    pub bwd_bulk_state_count: u64,
    bwd_bulk_packet_count_help: u64,
    bwd_bulk_start_help: Option<Instant>,
    bwd_bulk_size_help: u32,
    bwd_last_bulk_timestamp: Option<Instant>,

}

Represents a CIC Flow, encapsulating various metrics and states of a network flow.

This struct includes detailed information about both forward and backward flow, active and idle times, as well as subflows.

Fields
basic_flow: BasicFlow

The basic flow information.

sf_last_packet_timestamp: Option<Instant>

The timestamp of the last packet in the subflow.

sf_count: u32

The number of subflows.

start_active: Instant

The timestamp of the start of an active period.

end_active: Instant

The timestamp of the end of an active period.

active_count: u32

The number of active periods.

active_mean: f64

The mean of active periods.

active_std: f64

The standard deviation of active periods.

active_max: f64

The maximum active period.

active_min: f64

The minimum active period.

idle_count: u32

The number of idle periods.

idle_mean: f64

The mean of idle periods.

idle_std: f64

The standard deviation of idle periods.

idle_max: f64

The maximum idle period.

idle_min: f64

The minimum idle period.

fwd_init_win_bytes: u16

The initial window size of the forward flow.

fwd_act_data_pkt: u32

The number of data packets in the forward flow with more than one byte of data.

fwd_header_len_min: u32

The minimum header length of the forward flow.

fwd_last_timestamp: Option<Instant>

The timestamp of the last packet in the forward flow.

fwd_pkt_len_tot: u32

The total length of packets in the forward flow.

fwd_seg_len_tot: u32

The total length of transport layer segments in the forward flow (tcp/udp header + data).

fwd_pkt_len_max: u32

The maximum length of packets in the forward flow.

fwd_pkt_len_min: u32

The minimum length of packets in the forward flow.

fwd_pkt_len_mean: f32

The mean length of packets in the forward flow.

fwd_pkt_len_std: f32

The standard deviation of the length of packets in the forward flow.

fwd_iat_total: f64

The total inter-arrival time of packets in the forward flow.

fwd_iat_mean: f64

The mean inter-arrival time of packets in the forward flow.

fwd_iat_std: f64

The standard deviation of the inter-arrival time of packets in the forward flow.

fwd_iat_max: f64

The maximum inter-arrival time of packets in the forward flow.

fwd_iat_min: f64

The minimum inter-arrival time of packets in the forward flow.

fwd_header_length: u32

The total header length of the forward flow.

fwd_bulk_duration: f64

The total duration of bulk packets in the forward flow.

fwd_bulk_packet_count: u64

The number of bulk packets in the forward flow.

fwd_bulk_size_total: u32

The total size of bulk packets in the forward flow.

fwd_bulk_state_count: u64

The number of bulk states in the forward flow.

fwd_bulk_packet_count_help: u64

Helper variable for bulk packet count.

fwd_bulk_start_help: Option<Instant>

Helper variable for bulk start timestamp.

fwd_bulk_size_help: u32

Helper variable for bulk size.

fwd_last_bulk_timestamp: Option<Instant>

The timestamp of the last bulk packet in the forward flow.

bwd_init_win_bytes: u16

The initial window size of the backward flow.

bwd_last_timestamp: Option<Instant>

The timestamp of the last packet in the backward flow.

bwd_pkt_len_tot: u32

The total length of packets in the backward flow.

bwd_seg_len_tot: u32

The total length of transport layer segments in the backward flow (tcp/udp header + data).

bwd_pkt_len_max: u32

The maximum length of packets in the backward flow.

bwd_pkt_len_min: u32

The minimum length of packets in the backward flow.

bwd_pkt_len_mean: f32

The mean length of packets in the backward flow.

bwd_pkt_len_std: f32

The standard deviation of the length of packets in the backward flow.

bwd_iat_total: f64

The total inter-arrival time of packets in the backward flow.

bwd_iat_mean: f64

The mean inter-arrival time of packets in the backward flow.

bwd_iat_std: f64

The standard deviation of the inter-arrival time of packets in the backward flow.

bwd_iat_max: f64

The maximum inter-arrival time of packets in the backward flow.

bwd_iat_min: f64

The minimum inter-arrival time of packets in the backward flow.

bwd_header_length: u32

The total header length of the backward flow.

bwd_bulk_duration: f64

The total duration of bulk packets in the backward flow.

bwd_bulk_packet_count: u64

The number of bulk packets in the backward flow.

bwd_bulk_size_total: u32

The total size of bulk packets in the backward flow.

bwd_bulk_state_count: u64

The number of bulk states in the backward flow.

bwd_bulk_packet_count_help: u64

Helper variable for bulk packet count.

bwd_bulk_start_help: Option<Instant>

Helper variable for bulk start timestamp.

bwd_bulk_size_help: u32

Helper variable for bulk size.

bwd_last_bulk_timestamp: Option<Instant>

The timestamp of the last bulk packet in the backward flow.

Implementations

impl CicFlow

fn increase_fwd_header_length(&mut self, len: u32)

Increases the length of the forward header.

This method adds the specified length to the current forward header length. It’s used to accumulate the total size of the headers over multiple packets.

Arguments

len - The length to be added to the forward header length.

fn increase_bwd_header_length(&mut self, len: u32)

Increases the length of the backward header.

Similar to increase_fwd_header_length, this method accumulates the length of backward headers. It’s used in tracking the size of headers from the opposite direction of the flow.

Arguments

len - The length to be added to the backward header length.

fn update_fwd_header_len_min(&mut self, len: u32)

Updates the minimum length of the forward header.

This method updates the minimum forward header length if the provided length is smaller than the current minimum.

Arguments

len - The length to compare against the current minimum forward header length.

fn update_fwd_pkt_len_stats(&mut self, len: u32)

Updates statistics for the length of forward packets.

This method updates the maximum, minimum, total, mean, and standard deviation for the lengths of forward packets. It’s used for analyzing packet size distributions in the forward direction of the flow.

Arguments

len - The length of the new packet to be incorporated into the statistics.

fn update_bwd_pkt_len_stats(&mut self, len: u32)

Updates statistics for the length of backward packets.

Similar to update_fwd_pkt_len_stats, but for backward packets. It updates the statistical measures for packet sizes in the backward direction.

Arguments

len - The length of the new backward packet to be included in the stats.

fn update_fwd_iat_stats(&mut self, iat: f64)

Updates inter-arrival time (IAT) stats for forward packets.

This method updates the maximum, minimum, total, mean, and standard deviation of inter-arrival times between forward packets.

Arguments

iat - The inter-arrival time to be added to the statistics.

fn update_bwd_iat_stats(&mut self, iat: f64)

Updates inter-arrival time (IAT) stats for backward packets.

Similar to update_fwd_iat_stats, but focuses on the inter-arrival times of backward packets.

Arguments

iat - The backward inter-arrival time to be incorporated into the stats.

fn update_active_flow(&mut self, duration: f64)

Updates the statistics for active flow periods.

This method updates the count, mean, standard deviation, and max/min values for the durations of active flow periods.

Arguments

duration - The duration of the active period to be included in the stats.

fn update_idle_flow(&mut self, duration: f64)

Updates the statistics for idle flow periods.

Similar to update_active_flow, but for idle periods. This method updates the statistical measures related to the duration of idle periods in the flow.

Arguments

duration - The duration of the idle period to be added to the stats.

fn update_fwd_bulk_stats(&mut self, timestamp: &Instant, len: u32)

Updates the forward bulk statistics based on the incoming packet.

This method takes into account the timestamp and length of the packet to determine whether it’s part of an existing bulk transfer or the start of a new one. It updates various metrics related to bulk transfers in the forward direction.

Arguments

timestamp - The timestamp of the packet.

len - The length of the packet.

fn update_bwd_bulk_stats(&mut self, timestamp: &Instant, len: u32)

Updates the backward bulk statistics in a similar manner to update_fwd_bulk_stats.

It analyzes incoming packets in the backward direction and updates metrics related to bulk transfers based on the packet’s timestamp and length.

Arguments

timestamp - The timestamp of the packet.

len - The length of the packet.

fn update_subflows(&mut self, timestamp: &Instant)

Updates the subflow count based on the timestamp of the incoming packet.

This method increments the subflow count if the time since the last packet exceeds a threshold. It also updates the active and idle times of the flow based on the timestamp.

Arguments

timestamp - The timestamp of the packet.

fn update_active_idle_time(&mut self, timestamp: &Instant, threshold: f64)

Updates the active and idle time statistics of the flow.

Based on the timestamp and a specified threshold, this method determines whether the flow is active or idle, and updates the respective statistics. It is used for tracking how long the flow has been in each state.

Arguments

timestamp - The timestamp of the packet or event triggering the update.

threshold - The threshold in microseconds to determine state transitions between active and idle.

fn get_fwd_header_len_min(&self) -> u32

Retrieves the minimum length of the forward header.

Returns the minimum length observed for forward headers. If no forward headers have been recorded, it returns 0.

Returns

The minimum length of the forward header or 0 if not set.

pub fn get_flow_iat_std(&self) -> f64

Calculates the pooled standard deviation of inter-arrival times (IAT) for the flow.

This method considers both forward and backward packet inter-arrival times to compute a pooled variance, which is then square-rooted to get the standard deviation.

Returns

Pooled standard deviation of the flow’s IATs.

pub fn get_flow_iat_mean(&self) -> f64

Calculates the mean inter-arrival time (IAT) for the flow.

Averages the IATs of forward and backward packets to compute the overall mean IAT of the flow.

Returns

Mean inter-arrival time of the flow.

pub fn get_flow_iat_max(&self) -> f64

Retrieves the maximum inter-arrival time (IAT) observed in the flow.

Compares the maximum IATs in both directions (forward and backward) and returns the larger one.

Returns

Maximum inter-arrival time observed in the flow.

pub fn get_flow_iat_min(&self) -> f64

Retrieves the minimum inter-arrival time (IAT) observed in the flow.

Compares the minimum IATs in both directions and returns the smaller one, or 0 if not set.

Returns

Minimum inter-arrival time observed in the flow or 0 if not set.

pub fn get_fwd_iat_min(&self) -> f64

Retrieves the minimum IAT of packets in the forward flow.

Compares the minimum IAT to the max of the f64 type and returns the IAT if it is not the same as the max value.

Returns

The minimum IAT observed in the forward flow or 0 if not set.

pub fn get_bwd_iat_min(&self) -> f64

Retrieves the minimum IAT of packets in the backward flow.

Compares the minimum IAT to the max of the f64 type and returns the IAT if it is not the same as the max value.

Returns

The minimum IAT observed in the backward flow or 0 if not set.

pub fn get_flow_packet_length_min(&self) -> u32

Retrieves the minimum packet length in the flow, considering both forward and backward directions.

Compares the minimum packet lengths of forward and backward flows and returns the smaller value. Returns 0 if the minimum length is not set (indicated by u32::MAX).

Returns

Minimum packet length in the flow, or 0 if not set.

pub fn get_flow_packet_length_max(&self) -> u32

Retrieves the maximum packet length in the flow, considering both forward and backward directions.

Compares the maximum packet lengths of forward and backward flows and returns the larger value.

Returns

Maximum packet length in the flow.

pub fn get_fwd_packet_length_min(&self) -> u32

Retrieves the minimum packet length for forward packets.

Returns the minimum packet length observed in the forward direction. Returns 0 if the minimum length is not set (indicated by u32::MAX).

Returns

Minimum forward packet length, or 0 if not set.

pub fn get_bwd_packet_length_min(&self) -> u32

Retrieves the minimum packet length for backward packets.

Returns the minimum packet length observed in the backward direction. Returns 0 if the minimum length is not set (indicated by u32::MAX).

Returns

Minimum backward packet length, or 0 if not set.

pub fn get_flow_packet_length_mean(&self) -> f32

Calculates the mean packet length of the flow, averaging both forward and backward packet lengths.

The mean is computed by considering the lengths and counts of packets in both directions.

Returns

Mean packet length of the flow.

pub fn get_flow_segment_length_mean(&self) -> f32

Calculates the mean packet segment length of the flow, averaging both forward and backward packet segment lengths.

The mean is computed by considering the lengths and counts of packets in both directions.

Returns

Mean packet segment length of the flow.

pub fn get_flow_packet_length_variance(&self) -> f64

Calculates the variance of the packet lengths in the flow.

Computes the variance by considering the standard deviations of packet lengths in both forward and backward directions.

Returns

Variance of the flow’s packet lengths, or 0 if not enough data.

pub fn get_flow_packet_length_std(&self) -> f64

Retrieves the standard deviation of packet lengths in the flow.

Utilizes the calculated variance of packet lengths to compute the standard deviation.

Returns

Standard deviation of the flow’s packet lengths.

pub fn get_down_up_ratio(&self) -> f64

Calculates the down/up ratio of the flow.

Computes the ratio of the number of forward packets to backward packets. Returns 0 if there are no backward packets, to avoid division by zero.

Returns

The down/up ratio of the flow, or 0 if there are no backward packets.

pub fn get_fwd_segment_length_mean(&self) -> f32

Retrieves the mean segment length of forward packets.

Calculates the average segment length of forward packets by dividing the total segments length of forward packets by their count. Returns 0 if no forward packets are present.

Returns

Mean segment length of forward packets, or 0 if no forward packets are present.

pub fn get_bwd_segment_length_mean(&self) -> f32

Retrieves the mean segemnt length of backward packets.

Similar to get_fwd_segment_length_mean, but calculates the average segment length for backward packets. Returns 0 if no backward packets are present.

Returns

Mean segment length of backward packets, or 0 if no backward packets are present.

fn get_flow_bytes_s(&self) -> f64

Calculates the bytes per second rate of the flow.

Computes the total number of bytes (forward and backward) transferred in the flow and divides it by the total duration of the flow in seconds.

Returns

Bytes per second rate of the flow.

fn get_flow_packets_s(&self) -> f64

Calculates the packets per second rate of the flow.

Computes the total number of packets (forward and backward) in the flow and divides it by the total duration of the flow in seconds.

Returns

Packets per second rate of the flow.

pub fn get_fwd_packets_s(&self) -> f64

Calculates the forward packets per second rate of the flow.

Computes the number of forward packets in the flow and divides it by the total duration of the flow in seconds.

Returns

Forward packets per second rate of the flow.

pub fn get_bwd_packets_s(&self) -> f64

Calculates the backward packets per second rate of the flow.

Computes the number of backward packets in the flow and divides it by the total duration of the flow in seconds.

Returns

Backward packets per second rate of the flow.

pub fn get_fwd_bytes_bulk(&self) -> f64

Retrieves the average size of bulk transfers in the forward direction.

Calculates the mean size of bulk data transfers based on the total size and the number of bulk transfer states in the forward direction.

Returns

Average size of forward bulk transfers, or 0 if there are no bulk transfers.

pub fn get_fwd_packets_bulk(&self) -> f64

Retrieves the average number of packets in bulk transfers in the forward direction.

Calculates the mean number of packets in bulk transfers based on the total number and the count of bulk transfer states in the forward direction.

Returns

Average number of packets in forward bulk transfers, or 0 if there are no bulk transfers.

pub fn get_fwd_bulk_rate(&self) -> f64

Calculates the forward bulk rate.

Computes the rate of bulk data transfer in the forward direction by dividing the total size of forward bulk transfers by the total duration of these transfers in seconds.

Returns

Forward bulk data transfer rate in bytes per second, or 0 if there are no forward bulk transfers.

pub fn get_bwd_bytes_bulk(&self) -> f64

Retrieves the average size of bulk transfers in the backward direction.

Calculates the mean size of bulk data transfers based on the total size and the number of bulk transfer states in the backward direction.

Returns

Average size of backward bulk transfers, or 0 if there are no bulk transfers.

pub fn get_bwd_packets_bulk(&self) -> f64

Retrieves the average number of packets in bulk transfers in the backward direction.

Calculates the mean number of packets in bulk transfers based on the total number and the count of bulk transfer states in the backward direction.

Returns

Average number of packets in backward bulk transfers, or 0 if there are no bulk transfers.

pub fn get_bwd_bulk_rate(&self) -> f64

Calculates the backward bulk rate.

Computes the rate of bulk data transfer in the backward direction by dividing the total size of backward bulk transfers by the total duration of these transfers in seconds.

Returns

Backward bulk data transfer rate in bytes per second, or 0 if there are no backward bulk transfers.

pub fn get_sf_fwd_packets(&self) -> f64

Calculates the average number of forward packets per subflow.

Determines the mean number of forward packets across subflows.

Returns

Average number of forward packets per subflow, or 0 if there are no subflows.

pub fn get_sf_fwd_bytes(&self) -> f64

Calculates the average number of forward bytes per subflow.

Determines the mean number of bytes in the forward direction across subflows.

Returns

Average number of forward bytes per subflow, or 0 if there are no subflows.

pub fn get_sf_bwd_packets(&self) -> f64

Calculates the average number of backward packets per subflow.

Determines the mean number of backward packets across subflows.

Returns

Average number of backward packets per subflow, or 0 if there are no subflows.

pub fn get_sf_bwd_bytes(&self) -> f64

Calculates the average number of backward bytes per subflow.

Determines the mean number of bytes in the backward direction across subflows. It’s useful for understanding the data transfer characteristics in each identified subflow.

Returns

Average number of backward bytes per subflow, or 0 if there are no subflows.

pub fn get_active_min(&self) -> f64

Retrieves the minimum active time observed in the flow.

This function returns the shortest period of time in which the flow was active. If the minimum active time has never been updated (indicated by f64::MAX), it returns 0.0.

Returns

Minimum active time in microseconds, or 0.0 if not set.

pub fn get_idle_min(&self) -> f64

Retrieves the minimum idle time observed in the flow.

Similar to get_active_min, this function returns the shortest idle period in the flow. If the minimum idle time has never been updated (indicated by f64::MAX), it returns 0.0.

Returns

Minimum idle time in microseconds, or 0.0 if not set.