rustiflow

Module realtime

Structs

PacketGraphState 🔒

RealtimeEbpfCounters 🔒

RealtimeSourceStats 🔒

ShardDispatchBatch 🔒

Constants

DEFAULT_SHARD_BATCH_SIZE 🔒

DEFAULT_SHARD_QUEUE_CAPACITY 🔒

DEFAULT_SOURCE_DISPATCH_QUEUE_CAPACITY 🔒

Functions

bump_memlock_rlimit 🔒

compute_realtime_offset_us 🔒

compute_shard_index 🔒

create_pending_batches 🔒

ebpf_binary_path 🔒

elapsed_ns 🔒

enqueue_pending_batches 🔒

enqueue_shard_batch 🔒

handle_realtime

Starts the realtime processing of packets on the given interface. The function will return the number of packets dropped by the eBPF program.

labeled_ringbuf_sources 🔒

load_ebpf_ipv4 🔒

load_ebpf_ipv6 🔒

log_source_stats 🔒

queue_label 🔒

read_clock_us 🔒

read_env_usize 🔒

read_per_cpu_counter 🔒

run_source_dispatcher 🔒

spawn_source_dispatcher 🔒

take_ring_buf_maps 🔒

tc_attach_type_label 🔒