rustiflow
Module realtime

Functions

bump_memlock_rlimit 🔒

compute_shard_index 🔒

handle_realtime

Starts the realtime processing of packets on the given interface. The function will return the number of packets dropped by the eBPF program.

load_ebpf_ipv4 🔒

load_ebpf_ipv6 🔒